% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsim.R
\name{ac_fft}
\alias{ac_fft}
\title{Compute autocovariance or autocorrelation function via Wiener-Khinchin theorem
  using Fast Fourier Transform}
\usage{
ac_fft(x, demean = TRUE)
}
\arguments{
\item{x}{matrix with time (iteration number) along the rows and variables along the columns.}

\item{demean}{whether to subtract from each column its mean.}
}
\value{
A matrix of the same size as x with autocovariances at all lags from 1 to the number of rows.
}
\description{
Compute autocovariance or autocorrelation function via Wiener-Khinchin theorem
  using Fast Fourier Transform
}
\keyword{internal}
