% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMatrix.R
\name{anyNA,tabMatrix-method}
\alias{anyNA,tabMatrix-method}
\title{S4 method for generic 'anyNA' and signature 'tabMatrix'}
\usage{
\S4method{anyNA}{tabMatrix}(x, recursive = FALSE)
}
\arguments{
\item{x}{a tabMatrix object.}

\item{recursive}{not used.}
}
\value{
Whether the tabMatrix object contains missings or not.
}
\description{
S4 method for generic 'anyNA' and signature 'tabMatrix'
}
\keyword{internal}
