% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsim.R
\name{as.data.frame.dc_summary}
\alias{as.data.frame.dc_summary}
\title{Convert a dc_summary object to a data.frame}
\usage{
\method{as.data.frame}{dc_summary}(x, row.names, optional, ...)
}
\arguments{
\item{x}{a summary of a draws component object.}

\item{row.names}{not used.}

\item{optional}{not used.}

\item{...}{not used.}
}
\value{
The \code{dc_summary} object as a \code{data.frame}.
}
\description{
Convert a dc_summary object to a data.frame
}
