% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsim.R
\name{as.data.frame.mcdraws_summary}
\alias{as.data.frame.mcdraws_summary}
\title{Convert a mcdraws_summary object to a data.frame}
\usage{
\method{as.data.frame}{mcdraws_summary}(x, row.names, optional, ...)
}
\arguments{
\item{x}{a summary of an object of class \code{mcdraws}.}

\item{row.names}{not used.}

\item{optional}{not used.}

\item{...}{not used.}
}
\value{
The \code{mcdraws_summary} object as a \code{data.frame}.
}
\description{
Convert a mcdraws_summary object to a data.frame
}
