% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{bayesR2}
\alias{bayesR2}
\title{Compute a Bayesian measure of percentage of explained variance}
\usage{
bayesR2(x, summary = TRUE)
}
\arguments{
\item{x}{object of class mcdraws as created by \code{\link{MCMCsim}}.}

\item{summary}{if \code{TRUE} only a posterior summary is returned, otherwise
a simulation (\code{dc}) object.}
}
\value{
Posterior summary or full simulation draws of a Bayesian version of
 R-squared modelfit measure.
}
\description{
Compute a Bayesian measure of percentage of explained variance
}
\references{
A. Gelman, B. Goodrich, J. Gabry and A. Vehtari (2019).
   R-squared for Bayesian Regression Models.
   The American Statistician 73(3), 307-309.
}
