% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_binomial.R
\name{binomial_control}
\alias{binomial_control}
\title{Set computational options for the sampling algorithms}
\usage{
binomial_control(PG.approx = TRUE, PG.approx.m = -2L, probit.HaarPXDA = TRUE)
}
\arguments{
\item{PG.approx}{whether Polya-Gamma draws for logistic binomial models are
approximated by a hybrid gamma convolution approach. If not, \code{BayesLogit::rpg}
is used, which is exact for some values of the shape parameter.}

\item{PG.approx.m}{if \code{PG.approx=TRUE}, the number of explicit gamma draws in the
sum-of-gammas representation of the Polya-Gamma distribution. The remainder (infinite)
convolution is approximated by a single moment-matching gamma draw. Special values are:
\code{-2L} for a default choice depending on the value of the shape parameter
balancing performance and accuracy, \code{-1L} for a moment-matching normal approximation,
and \code{0L} for a moment-matching gamma approximation.}

\item{probit.HaarPXDA}{only used for binomial models with probit link. If supported,
\code{probit.HaarPXDA=TRUE} and the Haar PX-DA sandwich step is added to the Albert-Chib
data augmentation scheme for probit multilevel models. This will usually result in
a faster mixing MCMC algorithm. Currently supported when all coefficients are sampled
in a single Gibbs block and all coefficients' prior means equal zero.}
}
\value{
A list with computational options.
}
\description{
Set computational options for the sampling algorithms
}
