% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMVN_sampler.R
\name{cMVN_control}
\alias{cMVN_control}
\title{Set options for fast sampling from a degenerate multivariate normal distribution}
\usage{
cMVN_control(
  eps1 = sqrt(.Machine$double.eps),
  eps2 = sqrt(.Machine$double.eps),
  chol.control = chol_control()
)
}
\arguments{
\item{eps1}{scalar parameter to control numerical robustness against singularity of Q.}

\item{eps2}{scalar parameter associated with the constraint part to control numerical robustness.}

\item{chol.control}{options for Cholesky decomposition, see \code{\link{chol_control}}.}
}
\value{
A list of options used by the constrained MVN sampler.
}
\description{
Set options for fast sampling from a degenerate multivariate normal distribution
}
