% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{combine_chains}
\alias{combine_chains}
\title{Combine multiple mcdraws objects into a single one by combining their chains}
\usage{
combine_chains(...)
}
\arguments{
\item{...}{objects of class \code{mcdraws}.}
}
\value{
A combined object of class \code{mcdraws} where the number of stored
 chains equals the sum of the numbers of chains in the input objects.
}
\description{
This function can be used to combine the results of parallel simulations.
}
