% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{combine_iters}
\alias{combine_iters}
\title{Combine multiple mcdraws objects into a single one by combining their draws}
\usage{
combine_iters(...)
}
\arguments{
\item{...}{objects of class \code{mcdraws}}
}
\value{
A combined object of class \code{mcdraws} where the number of stored
 draws equals the sum of the numbers of draws in the input objects.
}
\description{
This function is used to combine the results of parallel posterior predictive simulations.
}
