% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_multi.R
\name{f_multi}
\alias{f_multi}
\title{Specify a multi-response sampling distribution}
\usage{
f_multi(fam.list, fam.ind)
}
\arguments{
\item{fam.list}{a named list of response types.}

\item{fam.ind}{a formula specifying a character or factor variable indicating
which observations are of each response type specified in \code{fam.list}.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a multi-response sampling distribution.
The current implementation requires the data to be in long format, disregarding
the possibility that observations may be on the same units.
}
