% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_poisson.R
\name{f_poisson}
\alias{f_poisson}
\title{Specify a Poisson sampling distribution}
\usage{
f_poisson(link = "log", control = poisson_control())
}
\arguments{
\item{link}{the name of a link function. Currently the only allowed
link function for the Poisson distribution is \code{"log"}.}

\item{control}{a list with computational options. These options can
be specified using function \code{\link{poisson_control}}.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a Poisson sampling distribution.
}
\examples{
\dontrun{
n <- 3000
m <- 25
dat <- data.frame(
  x = rnorm(n),
  g = factor(sample(1:m, n, replace=TRUE), levels=1:m)
)
v <- rnorm(m, sd=0.6)
dat$y <- rpois(n, lambda=exp(with(dat, 1 - 0.5*x + v[g])))

sampler <- create_sampler(
  y ~ reg(~ x, name="beta") +       # fixed effects
      gen(factor = ~ g, name="v"),  # random intercepts
  data=dat, family="poisson"
)
sim <- MCMCsim(sampler, store.all=TRUE)
compute_DIC(sim)
waic(sim)
summary(sim)

bayesplot::mcmc_recover_intervals(as.array(sim$beta), c(1, -0.5))
bayesplot::mcmc_recover_intervals(as.array(sim$v_sigma), 0.6)
yrep <- predict(sim, iters=sample(1:1000, 10))
bayesplot::pp_check(dat$y, as.matrix(yrep), bayesplot::ppc_dens_overlay)
}

}
