% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_student_t.R
\name{f_student_t}
\alias{f_student_t}
\title{Specify a Student-t sampling distribution}
\usage{
f_student_t(
  link = "identity",
  df.prior = pr_gamma(2, 0.1),
  control = student_t_control(),
  ...
)
}
\arguments{
\item{link}{the name of a link function. Currently the only allowed link function
for a Student-t distribution is \code{"identity"}.}

\item{df.prior}{prior specification for the (scalar) number of degrees of
freedom of the Studen-t sampling distribution. Currently allowed priors are
gamma priors specified using \code{\link{pr_gamma}}, or a degenerate prior
fixing the degrees of freedom parameter at a fixed value.}

\item{control}{sampling options for the degrees of freedom parameter
(if not fixed).}

\item{...}{further arguments passed to \code{\link{f_gaussian}}.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a Student-t sampling distribution.
}
