% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gaussian.R
\name{gaussian_control}
\alias{gaussian_control}
\title{Set computational options for the sampling algorithms}
\usage{
gaussian_control(block.V = TRUE)
}
\arguments{
\item{block.V}{if \code{TRUE}, the default, all coefficients of \code{reg}
and \code{gen} components in a variance model formula are sampled in a
single block. Alternatively, a list of character vectors with names of
model components whose coefficients should be sampled together in blocks.}
}
\value{
A list with computational options.
}
\description{
Set computational options for the sampling algorithms
}
