% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{Matrix-methods}
\alias{Matrix-methods}
\alias{\%*\%,ddiMatrix,matrix-method}
\alias{\%*\%,dgCMatrix,matrix-method}
\alias{\%*\%,dsCMatrix,matrix-method}
\alias{\%*\%,tabMatrix,matrix-method}
\alias{\%*\%,matrix,tabMatrix-method}
\alias{\%*\%,tabMatrix,numLike-method}
\alias{\%*\%,ddiMatrix,dgCMatrix-method}
\alias{\%*\%,ddiMatrix,tabMatrix-method}
\alias{\%*\%,CsparseMatrix,tabMatrix-method}
\alias{\%*\%,tabMatrix,CsparseMatrix-method}
\alias{tcrossprod,matrix,dgCMatrix-method}
\alias{tcrossprod,matrix,tabMatrix-method}
\alias{tcrossprod,matrix,ddiMatrix-method}
\alias{crossprod,tabMatrix,missing-method}
\alias{crossprod,tabMatrix,matrix-method}
\alias{crossprod,tabMatrix,dgCMatrix-method}
\alias{crossprod,tabMatrix,tabMatrix-method}
\alias{crossprod,dgCMatrix,matrix-method}
\title{S4 methods for products of matrix objects}
\usage{
\S4method{\%*\%}{ddiMatrix,matrix}(x, y)

\S4method{\%*\%}{dgCMatrix,matrix}(x, y)

\S4method{\%*\%}{dsCMatrix,matrix}(x, y)

\S4method{\%*\%}{tabMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,tabMatrix}(x, y)

\S4method{\%*\%}{tabMatrix,numLike}(x, y)

\S4method{\%*\%}{ddiMatrix,dgCMatrix}(x, y)

\S4method{\%*\%}{ddiMatrix,tabMatrix}(x, y)

\S4method{\%*\%}{CsparseMatrix,tabMatrix}(x, y)

\S4method{\%*\%}{tabMatrix,CsparseMatrix}(x, y)

\S4method{tcrossprod}{matrix,dgCMatrix}(x, y)

\S4method{tcrossprod}{matrix,tabMatrix}(x, y)

\S4method{tcrossprod}{matrix,ddiMatrix}(x, y)

\S4method{crossprod}{tabMatrix,missing}(x, y)

\S4method{crossprod}{tabMatrix,matrix}(x, y)

\S4method{crossprod}{tabMatrix,dgCMatrix}(x, y)

\S4method{crossprod}{tabMatrix,tabMatrix}(x, y)

\S4method{crossprod}{dgCMatrix,matrix}(x, y)
}
\arguments{
\item{x}{a matrix object.}

\item{y}{a matrix object.}
}
\value{
A matrix object. In case one of the arguments is a regular (dense) \code{matrix}
 the result is a \code{matrix} as well.
}
\description{
Several methods for products of matrix objects. Here a matrix object can be
an ordinary (dense) \code{matrix} or a (sparse) \code{\link[Matrix]{Matrix}} of class
\code{\link[Matrix]{ddiMatrix-class}}, \code{tabMatrix},
\code{\link[Matrix]{dgCMatrix-class}}, or \code{\link[Matrix]{dsCMatrix-class}}.
The return types are restricted to \code{matrix} or \code{numeric} in case of dense objects, and
\code{dgCMatrix}, \code{dsCMatrix}, \code{ddiMatrix} or \code{tabMatrix} in case of sparse objects.
}
\keyword{internal}
