% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcsae.R
\name{mcmcsae-package}
\alias{mcmcsae-package}
\alias{mcmcsae}
\title{Markov Chain Monte Carlo Small Area Estimation}
\description{
Fit multi-level models with possibly correlated random effects using MCMC.
}
\details{
Functions to fit multi-level models with Gaussian, binomial, multinomial,
negative binomial or Poisson likelihoods using MCMC. Models with a linear predictor
consisting of various possibly correlated random effects are supported, allowing
flexible modelling of temporal, spatial or other kinds of dependence structures.
For Gaussian models the variance can be modelled too. By modelling variances
at the unit level the marginal distribution can be changed to a Student-t or Laplace
distribution, which may account better for outliers.
The package has been developed with applications to small area estimation
in official statistics in mind. The posterior samples for the model
parameters can be passed to a prediction function to generate samples from
the posterior predictive distribution for user-defined quantities such as
finite population domain means. For model assessment, posterior predictive
checks and DIC/WAIC criteria can easily be computed.
}
