% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcmcsae_example}
\alias{mcmcsae_example}
\title{Generate artificial data according to an additive spatio-temporal model}
\usage{
mcmcsae_example(n = 100L, family = "gaussian")
}
\arguments{
\item{n}{the size of the generated dataset.}

\item{family}{sampling distribution family, see \code{\link{create_sampler}}.}
}
\value{
A \code{list} containing the generated dataset, the values of the model
  parameters, and the model specification as a formula.
}
\description{
This function is used to generate data for several examples.
}
\examples{
\donttest{
ex <- mcmcsae_example()
str(ex)
}

}
