% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_negbinomial.R
\name{negbin_control}
\alias{negbin_control}
\title{Set computational options for the sampling algorithms}
\usage{
negbin_control(CRT.approx.m = 20L)
}
\arguments{
\item{CRT.approx.m}{scalar integer specifying the degree of approximation to sampling
from a Chinese Restaurant Table distribution. The approximation is based on Le Cam's theorem.
Larger values yield a slower but more accurate sampler.}
}
\value{
A list with computational options for the sampling algorithm.
}
\description{
Set computational options for the sampling algorithms
}
