% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_negbinomial.R
\name{negbinomial_control}
\alias{negbinomial_control}
\title{Set computational options for the sampling algorithms}
\usage{
negbinomial_control(PG.approx = TRUE, PG.approx.m = -2L, CRT.approx.m = 20L)
}
\arguments{
\item{PG.approx}{whether Polya-Gamma draws for logistic binomial models are
approximated by a hybrid gamma convolution approach. If not, \code{BayesLogit::rpg}
is used, which is exact for some values of the shape parameter.}

\item{PG.approx.m}{if \code{PG.approx=TRUE}, the number of explicit gamma draws in the
sum-of-gammas representation of the Polya-Gamma distribution. The remainder (infinite)
convolution is approximated by a single moment-matching gamma draw. Special values are:
\code{-2L} for a default choice depending on the value of the shape parameter
balancing performance and accuracy, \code{-1L} for a moment-matching normal approximation,
and \code{0L} for a moment-matching gamma approximation.}

\item{CRT.approx.m}{scalar integer specifying the degree of approximation to sampling
from a Chinese Restaurant Table distribution. The approximation is based on Le Cam's theorem.
Larger values yield a slower but more accurate sampler.}
}
\value{
A list with computational options for the sampling algorithm.
}
\description{
Set computational options for the sampling algorithms
}
