% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.dc}
\alias{plot.dc}
\title{Trace, density and autocorrelation plots for (parameters of a) draws
component (dc) object}
\usage{
\method{plot}{dc}(x, nrows, ncols, ask = FALSE, ...)
}
\arguments{
\item{x}{a draws component object.}

\item{nrows}{number of rows in plot layout.}

\item{ncols}{number of columns in plot layout.}

\item{ask}{ask before plotting the next page; default is \code{FALSE}.}

\item{...}{arguments passed to \code{\link[stats]{density}}.}
}
\description{
Trace, density and autocorrelation plots for (parameters of a) draws
component (dc) object
}
\examples{
\donttest{
ex <- mcmcsae_example(n=50)
sampler <- create_sampler(ex$model, data=ex$dat)
sim <- MCMCsim(sampler, store.all=TRUE)
plot(sim$u)
}

}
