% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_beta}
\alias{pr_beta}
\title{Create an object representing beta prior distributions}
\usage{
pr_beta(a = 1, b = 1)
}
\arguments{
\item{a}{positive shape parameter.}

\item{b}{positive shape parameter.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing beta prior distributions
}
