% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_exp}
\alias{pr_exp}
\title{Create an object representing exponential prior distributions}
\usage{
pr_exp(scale = 1)
}
\arguments{
\item{scale}{scalar or vector scale parameter.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing exponential prior distributions
}
