% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_fixed}
\alias{pr_fixed}
\title{Create an object representing a degenerate prior fixing a
parameter (vector) to a fixed value}
\usage{
pr_fixed(value = 1)
}
\arguments{
\item{value}{scalar or vector value parameter.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing a degenerate prior fixing a
parameter (vector) to a fixed value
}
