% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_gamma}
\alias{pr_gamma}
\title{Create an object representing gamma prior distributions}
\usage{
pr_gamma(shape = 1, rate = 1)
}
\arguments{
\item{shape}{scalar or vector shape parameter.}

\item{rate}{scalar or vector rate, i.e. inverse scale, parameter.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing gamma prior distributions
}
