% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_gig}
\alias{pr_gig}
\title{Create an object representing Generalized Inverse Gaussian (GIG) prior distributions}
\usage{
pr_gig(a, b, p)
}
\arguments{
\item{a}{scalar or vector parameter.}

\item{b}{scalar or vector parameter.}

\item{p}{scalar or vector parameter.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing Generalized Inverse Gaussian (GIG) prior distributions
}
