% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_normal}
\alias{pr_normal}
\title{Create an object representing a possibly multivariate normal prior distribution}
\usage{
pr_normal(mean = 0, precision = 0, labels = NULL)
}
\arguments{
\item{mean}{scalar or vector mean parameter.}

\item{precision}{scalar, vector or matrix precision parameter.}

\item{labels}{optional character vector with coefficient labels. If specified,
it should have the same length as at least one of \code{mean} and \code{precision},
and in that case the normal prior with these parameters is assigned to these coefficients,
while any coefficients not present in labels will be assigned a non-informative
prior with mean 0 and precision 0.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing a possibly multivariate normal prior distribution
}
