% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_truncnormal}
\alias{pr_truncnormal}
\title{Create an object representing truncated normal prior distributions}
\usage{
pr_truncnormal(mean = 0, precision = 1, lower = 0, upper = Inf)
}
\arguments{
\item{mean}{scalar or vector mean parameter.}

\item{precision}{scalar, vector or matrix precision parameter.}

\item{lower}{lower limit of the truncated interval.}

\item{upper}{lower limit of the truncated interval.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing truncated normal prior distributions
}
