% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_unif}
\alias{pr_unif}
\title{Create an object representing uniform prior distributions}
\usage{
pr_unif(min = 0, max = 1)
}
\arguments{
\item{min}{lower limit.}

\item{max}{upper limit.}
}
\value{
An environment representing the specified prior, for internal use.
}
\description{
Create an object representing uniform prior distributions
}
