% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Set up a cluster for parallel computing}
\usage{
setup_cluster(n.cores = NULL, seed = NULL, export = NULL)
}
\arguments{
\item{n.cores}{the number of cpu cores to use.}

\item{seed}{optional random seed for reproducibility.}

\item{export}{a character vector with names of objects to export to the workers.}
}
\value{
An object representing the cluster.
}
\description{
The cluster is set up for a number of workers by loading the \pkg{mcmcsae}
package and setting up independent RNG streams.
}
