% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stop a cluster}
\usage{
stop_cluster(cl)
}
\arguments{
\item{cl}{the cluster object.}
}
\value{
\code{NULL}.
}
\description{
Stop a cluster set up by \code{\link{setup_cluster}}.
}
