% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_student_t.R
\name{student_t_control}
\alias{student_t_control}
\title{Set computational options for the sampling algorithms}
\usage{
student_t_control(proposal = "RW", MHscale = 1, adapt = TRUE)
}
\arguments{
\item{proposal}{"RW" for random walk Metropolis-Hastings or "mala"
for Metropolis-adjusted Langevin.}

\item{MHscale}{(starting) scale of Metropolis-Hastings update.}

\item{adapt}{whether to adapt the scale of the proposal distribution
during burnin to achieve better acceptance rates.}
}
\value{
A list with computational options.
}
\description{
Set computational options for the sampling algorithms
}
