% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMatrix.R
\name{tabMatrix-indexing}
\alias{tabMatrix-indexing}
\alias{[,tabMatrix,index,missing,logical-method}
\alias{[,tabMatrix,index,missing,missing-method}
\alias{[,tabMatrix,missing,index,logical-method}
\alias{[,tabMatrix,missing,index,missing-method}
\title{S4 method for row and column subsetting a 'tabMatrix'}
\usage{
\S4method{[}{tabMatrix,index,missing,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{tabMatrix,index,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{tabMatrix,missing,index,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{tabMatrix,missing,index,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a tabMatrix object.}

\item{i}{integer vector indicating the rows to select. Not used for column subsetting.}

\item{j}{integer vector indicating the columns to select. Not used for row subsetting.}

\item{...}{not used.}

\item{drop}{whether to return a vector in case of a single selected row or column.}
}
\value{
The selected rows/columns as a tabMatrix or a vector.
}
\description{
S4 method for row and column subsetting a 'tabMatrix'
}
\keyword{internal}
