% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_binomial.R
\name{f_binomial}
\alias{f_binomial}
\title{Specify a binomial sampling distribution}
\usage{
f_binomial(link = c("logit", "probit"), n.trial = NULL)
}
\arguments{
\item{link}{the name of a link function. Currently the only allowed link functions
for the binomial distribution are \code{"logit"} (default) and \code{"probit"}.}

\item{n.trial}{the number of binomial trials. This can be specified either as
a formula for a variable number of trials, or as a scalar value for a common
number of trials for all units.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a binomial sampling distribution. This
includes the special case of binary (Bernoulli) data.
}
