% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_multinomial.R
\name{f_multinomial}
\alias{f_multinomial}
\title{Specify a multinomial sampling distribution}
\usage{
f_multinomial(link = "logit", n.trial = NULL, K = NULL)
}
\arguments{
\item{link}{the name of a link function. Currently the only allowed link function
for the multinomial distribution is \code{"logit"}.}

\item{n.trial}{the number of multinomial trials. This can be specified either as
a formula for a variable number of trials, or as a scalar value for a common
number of trials for all units.}

\item{K}{number of categories for multinomial model; only used for prior predictive sampling.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a multinomial sampling distribution. This
includes the special case of categorical (multinoulli) data.
}
