% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_poisson.R
\name{f_poisson}
\alias{f_poisson}
\title{Specify a Poisson sampling distribution}
\usage{
f_poisson(link = "log", control = poisson_control())
}
\arguments{
\item{link}{the name of a link function. Currently the only allowed
link function for the Poisson distribution is \code{"log"}.}

\item{control}{a list with computational options. These options can
be specified using function \code{\link{poisson_control}}.}
}
\value{
A family object.
}
\description{
This function can be used in the \code{family} argument of \code{\link{create_sampler}}
or \code{\link{generate_data}} to specify a Poisson sampling distribution.
}
