% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_poisson.R
\name{poisson_control}
\alias{poisson_control}
\title{Set computational options for the sampling algorithms}
\usage{
poisson_control(nb.shape = 100)
}
\arguments{
\item{nb.shape}{shape parameter of the negative binomial distribution used
internally to approximate the Poisson distribution. This should be set to a relatively
large value (default is 100), corresponding to negligible overdispersion, to obtain a
good approximation to the Poisson sampling distribution. However, note that very large
values may cause slow MCMC exploration of the posterior distribution.}
}
\value{
A list with computational options for the sampling algorithm.
}
\description{
Set computational options for the sampling algorithms
}
